refactor(Video): change name for TodaysVideo
